vm/requirements.txt
torch==2.0.1
diffusers==0.17.1
transformers==4.30.2
accelerate==0.20.3
pillow==9.5.0
qrcode==7.4.2
numpy